separate hpp template from other functions interface
